Match Analysis in Django
When a user searches their in-game name, the back-end will (very slowly)
find and store data from each of their last 15 matches. Once stored, the
front-end displays a list of all matches in database, and upon clicking
the id of a match, the user is given their performance vs. the average,
and their final build.

Currently my efforts are in presenting the build in an intuitive way,
and providing timeline support for build display, as well as comparison
with a higher-tier player with timeline support.